requirements.txt
aiohttp==3.7.4.post0
antlr4-python3-runtime==4.9.2
appdirs==1.4.4
arrow==1.1.0
asn1crypto==1.4.0
astroid==2.6.6
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.5
azure-common==1.1.27
azure-core==1.13.0
azure-storage-blob==12.8.1
bcrypt==3.2.0
binaryornot==0.4.4
black==21.5b1
botocore>=1.20.56
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==8.0.1
common==0.1.0
cookiecutter==1.6.0
coverage==5.5
flake8==3.8.4
future==0.18.2
greenlet==1.1.0
idna==2.8
iniconfig==1.1.1
isodate==0.6.0
isort==5.9.3
jedi==0.17.2
Jinja2==3.0.1
jinja2-time==0.2.0
jmespath==0.10.0
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
msgpack==1.0.2
msrest==0.6.21
multidict==5.1.0
mypy==0.910
mypy-extensions==0.4.3
neo4j==4.3.4
neovim==0.3.1
numpy==1.20.3
oauthlib==3.1.0
oscrypto==1.2.1
Owlready==0.3.1
packaging==20.9
pandas==1.2.4
paramiko==2.7.2
parso==0.7.1
pathspec==0.8.1
pluggy==0.13.1
poyo==0.5.0
psycopg2==2.8.6
py==1.10.0
pycodestyle==2.6.0
pycparser==2.20
pycryptodomex==3.10.1
pydantic==1.8.2
pydocstyle==6.1.1
pyflakes==2.2.0
pylint==2.9.6
pyls-mypy==0.1.8
PyNaCl==1.4.0
pynvim==0.4.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.0
python-dateutil==2.8.1
python-http-client==3.3.2
python-json-logger==0.1.11
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
PyYAML==5.4.1
regex==2020.11.13
requests-oauthlib==1.3.0
rope==0.19.0
s3transfer==0.4.2
sendgrid==6.1.0
six==1.16.0
snowballstemmer==2.1.0
snowflake-connector-python==2.3.1
sourcery==1.2.7
sshtunnel==0.1.5
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
ujson==4.0.2
whichcraft==0.6.1
wrapt==1.12.1
yapf==0.31.0
yarl==1.6.3
